realtime_dashboard/requirements.txt
pandas
kafka-python
colorcet
bokeh